requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
bsedata==0.4.0
casparser==0.4.7
casparser-isin==2021.6.1
certifi==2020.4.5.1
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cryptography==3.4.4
cssselect==1.1.0
daphne==3.0.1
dateutils==0.6.8
Django==2.2.24
django-solo==1.1.5
djangorestframework==3.11.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
google-currency==1.0.9
huey==2.3.0
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
isort==5.6.4
jieba3k==0.35.1
joblib==1.0.1
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
mftool==1.8
multitasking==0.0.9
newspaper3k==0.2.8
nltk==3.5
nsetools==1.0.11
numpy==1.19.5
packaging==20.9
pandas==1.2.5
pdfminer.six==20201018
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
pyotp==2.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2019.3
PyYAML==5.4.1
rapidfuzz==1.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
rich==10.4.0
robin-stocks==2.0.3
rope==0.18.0
scipy==1.6.0
selenium==3.141.0
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.14.0
sortedcontainers==2.4.0
soupsieve==2.0.1
sqlparse==0.3.1
tika==1.24
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.59.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
yfinance==0.1.63
zope.interface==5.2.0
